refactor: move ResultType into assets/types
