+++ webapp/src/components/ActivityPage/Transaction/Transaction.tsx
@@ -22,12 +22,12 @@ import {
   CANCEL_BID_SUCCESS
 } from '../../../modules/bid/actions'
 import { locations } from '../../../modules/routing/locations'
-import { ResultType } from '../../../modules/routing/types'
 import { getContract } from '../../../modules/contract/utils'
 import { AssetProvider } from '../../AssetProvider'
 import { Mana } from '../../Mana'
 import { TransactionDetail } from './TransactionDetail'
 import { Props } from './Transaction.types'
+import { ResultType } from '../../../modules/asset/types'
 
 const Transaction = (props: Props) => {
   const { tx } = props
