+++ webapp/src/components/AssetPage/AssetPage.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { Page } from 'decentraland-ui'
 
-import { Asset, ResultType } from '../../modules/routing/types'
+import { Asset } from '../../modules/routing/types'
+import { ResultType } from '../../modules/asset/types'
 import { Navbar } from '../Navbar'
 import { Footer } from '../Footer'
 import { Navigation } from '../Navigation'
