+++ webapp/src/components/AssetPage/Owner/Owner.tsx
@@ -3,14 +3,11 @@ import { Link } from 'react-router-dom'
 import { Profile } from 'decentraland-dapps/dist/containers'
 import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 import { locations } from '../../../modules/routing/locations'
-import {
-  NFTBrowseOptions,
-  ResultType,
-  Section
-} from '../../../modules/routing/types'
+import { NFTBrowseOptions, Section } from '../../../modules/routing/types'
 import { VendorName } from '../../../modules/vendor'
 import { Props } from './Owner.types'
 import './Owner.css'
+import { ResultType } from '../../../modules/asset/types'
 
 const Owner = (props: Props) => {
   const { asset } = props
