+++ webapp/src/components/AssetPage/WearableHighlights/WearableHighlights.types.ts
@@ -1,7 +1,7 @@
 import { Dispatch } from 'redux'
 import { CallHistoryMethodAction } from 'connected-react-router'
 import { NFT } from '@dcl/schemas'
-import { ResultType } from '../../../modules/routing/types'
+import { ResultType } from '../../../modules/asset/types'
 
 export type Props = {
   type: ResultType
