+++ webapp/src/components/AssetProviderPage/AssetProviderPage.container.ts
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 
 import { RootState } from '../../modules/reducer'
-import { ResultType } from '../../modules/routing/types'
+import { ResultType } from '../../modules/asset/types'
 import { isConnecting } from '../../modules/wallet/selectors'
 import { MapStateProps, OwnProps } from './AssetProviderPage.types'
 import AssetProviderPage from './AssetProviderPage'
