+++ webapp/src/components/Bid/Bid.tsx
@@ -5,12 +5,12 @@ import { Profile } from 'decentraland-dapps/dist/containers'
 import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 import { locations } from '../../modules/routing/locations'
 import { addressEquals } from '../../modules/wallet/utils'
+import { ResultType } from '../../modules/asset/types'
 import { AssetProvider } from '../AssetProvider'
 import { AssetImage } from '../AssetImage'
 import { Mana } from '../Mana'
 import { AcceptButton } from './AcceptButton'
 import { WarningMessage } from './WarningMessage'
-import { ResultType } from '../../modules/routing/types'
 import { formatMANA } from '../../lib/mana'
 import { formatDistanceToNow } from '../../lib/date'
 import { Props } from './Bid.types'
