+++ webapp/src/components/BidPage/BidPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Page } from 'decentraland-ui'
-import { ResultType } from '../../modules/routing/types'
+import { ResultType } from '../../modules/asset/types'
 import { Navbar } from '../Navbar'
 import { Footer } from '../Footer'
 import { Wallet } from '../Wallet'
