+++ webapp/src/components/BuyPage/BuyPage.tsx
@@ -10,7 +10,8 @@ import { AssetProviderPage } from '../AssetProviderPage'
 import { NFT } from '../../modules/nft/types'
 import { isOwnedBy } from '../../modules/nft/utils'
 import { Order } from '../../modules/order/types'
-import { Asset, ResultType } from '../../modules/routing/types'
+import { Asset } from '../../modules/routing/types'
+import { ResultType } from '../../modules/asset/types'
 import { BuyNFTModal } from './BuyNFTModal'
 import { BuyItemModal } from './BuyItemModal'
 import { Props } from './BuyPage.types'
