+++ webapp/src/components/CancelSalePage/CancelSalePage.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { t, T } from 'decentraland-dapps/dist/modules/translation/utils'
 import { Page, Header, Button } from 'decentraland-ui'
 
+import { ResultType } from '../../modules/asset/types'
 import { locations } from '../../modules/routing/locations'
-import { ResultType } from '../../modules/routing/types'
 import { getAssetName } from '../../modules/nft/utils'
 import { formatMANA } from '../../lib/mana'
 import { Navbar } from '../Navbar'
