+++ webapp/src/components/NFTList/NFTList.types.ts
@@ -4,7 +4,7 @@ import { Item } from '@dcl/schemas'
 import { NFT } from '../../modules/nft/types'
 import { VendorName } from '../../modules/vendor/types'
 import { browseNFTs, BrowseNFTsAction } from '../../modules/routing/actions'
-import { ResultType } from '../../modules/routing/types'
+import { ResultType } from '../../modules/asset/types'
 
 export type Props = {
   vendor: VendorName
