+++ webapp/src/components/TransferPage/TransferPage.tsx
@@ -7,8 +7,8 @@ import { Wallet } from '../Wallet'
 import { AssetProviderPage } from '../AssetProviderPage'
 import { AssetAction } from '../AssetAction'
 import { locations } from '../../modules/routing/locations'
-import { ResultType } from '../../modules/routing/types'
 import { getAssetName, isOwnedBy } from '../../modules/nft/utils'
+import { ResultType } from '../../modules/asset/types'
 import { Props } from './TransferPage.types'
 import './TransferPage.css'
