+++ webapp/src/modules/asset/types.ts
@@ -1,0 +1,4 @@
+export enum ResultType {
+  ITEM = 'item',
+  NFT = 'nft'
+}
